특강_Serving2/fast_api/requirements.txt
timm
transformers==4.15.0
fairscale
# pycocoevalcap
# torch
# torchvision
Pillow
scipy
git+https://github.com/openai/CLIP.git
fastapi
uvicon[standard]
pydantic
python-multipart